iscsi-scst: Change (signed) into (signed int)
git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@6975 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ iscsi-scst/kernel/iscsi.c
@@ -1817,7 +1817,7 @@ static int nop_out_start(struct iscsi_cmnd *cmnd)
 			}
 
 			/* We already checked it in check_segment_length() */
-			sBUG_ON(cmnd->sg_cnt > (signed)ISCSI_CONN_IOV_MAX);
+			sBUG_ON(cmnd->sg_cnt > (signed int)ISCSI_CONN_IOV_MAX);
 
 			cmnd->own_sg = 1;
 			cmnd->bufflen = size;
@@ -1836,7 +1836,7 @@ static int nop_out_start(struct iscsi_cmnd *cmnd)
 			 * accesses to dummy_page, since for ISCSI_RESERVED_TAG
 			 * the data only read and then discarded.
 			 */
-			for (i = 0; i < (signed)ISCSI_CONN_IOV_MAX; i++) {
+			for (i = 0; i < (signed int)ISCSI_CONN_IOV_MAX; i++) {
 				conn->read_iov[i].iov_base =
 					page_address(dummy_page);
 				tmp = min_t(u32, size, PAGE_SIZE);
